[mms] SECURITY: Don't leak PDO DSN authentication/connection information in exception messages.
Exception messages may be displayed (desired or not) to the end-user.

Not sure if we want to display a lesser version of the DSN ... I will
leave that decision up to someone else.

Also, log the error or else there is a chance that it won't be logged at
the user level.